Split typography into fontSize and heading Move removeREM into own util function
